ai-service/requirements.txt
flask
flask_cors
opencv-python
imagehash
pillow
numpy
